Update xiaomi-aqara-button.groovy
bspranger - firstly thank you so much for putting this out there, I have found it so useful. However I bought my Aqara Button in Jan 2020 and am running Samsung SmartThings Hub v3 with the new app not the classic app and the original code did not recognize the double press (and possibly the held - it's has been a long and steep learning curve to sort this for me I have never looked at anything to do with SmartThings as I have only had the hub and the button for a week!) button. Anyway the docs for the capability of a button for the new SmartThings App show that the event type has to be double - not double-clicked or pushed or Double Pressed but just double in order for SmartThings to recognise the event and act upon it. I have amended the code and my have it here and my button (Model WXKG11LM (new revision)) now recognises single click, double click and held. I hope this helps others and once again thanks for your DH (Device Handler for total noobs like me!)